+++ web/app/(frontend)/(logged-in)/layout.tsx
@@ -7,7 +7,7 @@ import "../globals.css"
 import { TailwindIndicator } from "@/components/tailwind-indicator"
 import { ThemeProvider } from "@/components/theme-provider"
 import { Footer } from "../components/footer"
-import { Nav } from "../components/nav"
+import { Nav } from "./components/nav"
 import { SpeedInsights } from "@vercel/speed-insights/next"
 
 export const metadata: Metadata = {
